SBERDOMA-933 move analytics to separate dir
