+++ apps/condo/pages/analytics/detail/report_by_tickets/pdf.tsx
